requirements.txt
absl-py==2.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.26.82
botocore==1.29.82
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.0.1
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.3.3
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.34.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kaggle==1.6.11
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.2
matplotlib==3.8.4
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
multidict==6.0.4
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
openai==0.26.5
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.6.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.13.0
selenium==4.9.1
Send2Trash==1.8.3
shap==0.44.1
six==1.16.0
slicer==0.0.7
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.2
trio==0.22.0
trio-websocket==0.10.2
twilio==7.16.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
